Fix header adding icons and login-logout nav items
