+++ apps/about/templates/about/index.html
@@ -10,7 +10,7 @@
 <p class="subtitle is-3">{{ pythoncanarias.name }}</p>
 
 <div class="box has-text-centered">
-  <a href="{% url 'members:join' %}" class="link button is-link is-medium is-outlined">Forma parte de la asociación</a>
+  <a href="{% url 'about:join' %}" class="link button is-link is-medium is-outlined">Forma parte de la asociación</a>
 </div>
 
 <article class="message">
@@ -26,7 +26,7 @@
 
     <p>Te recomendamos que no dejes de <a href="{% url 'about:history' %}">repasar la historia de Python Canarias</a>.</p>
 
-    <p>Si quieres apoyar nuestra labor para que podamos abordar proyectos cada vez más ambiciosos, <a href="{% url 'members:join' %}">hazte socio/a!</a> Estamos deseando escuchar tus propuestas.</p>
+    <p>Si quieres apoyar nuestra labor para que podamos abordar proyectos cada vez más ambiciosos, <a href="{% url 'about:join' %}">hazte socio/a!</a> Estamos deseando escuchar tus propuestas.</p>
   </div>
 </article>
