+++ apps/about/templates/about/join.html
@@ -1,8 +1,8 @@
-{% extends "members/base.html" %}
+{% extends "about/base.html" %}
 {% load utils %}
 {% load static %}
 
-{% block title %}Socios/as - {{ block.super }}{% endblock %}
+{% block title %}Únete - {{ block.super }}{% endblock %}
 
 {% block content %}
