+++ apps/commons/templates/header.html
@@ -15,15 +15,29 @@
 
   <div class="navbar-menu" id="navMenu">
     <div class="navbar-end has-text-centered">
-      <a class="navbar-item {% is_active 'about:index' %} {% is_active 'members:homepage' %}" href="{% url 'about:index' %}">
+      <a class="navbar-item {% is_active 'about:index' %}" href="{% url 'about:index' %}">
+        <i class="fas fa-user-friends"></i>
         Asociación
       </a>
       <a class="navbar-item {% is_active 'events:index' %}" href="{% url 'events:index' %}">
+        <i class="fas fa-birthday-cake"></i>
         Eventos
       </a>
       <a class="navbar-item {% is_active 'jobs:index' %}" href="{% url 'jobs:index' %}">
+        <i class="fas fa-laptop-code"></i>
         Empleo
       </a>
+
+      {% if request.user.is_authenticated %}
+        <a class="navbar-item {% is_active 'members:homepage' %}" href="{% url 'members:profile' %}">
+          <i class="fas fa-user-cog"></i>
+          {{ request.user.username }}
+        </a>
+      {% else %}
+        <a class="navbar-item" href="{% url 'members:login' %}">
+          <i class="fas fa-sign-in-alt"></i>
+        </a>
+      {% endif %}
     </div>
   </div>
 </nav>
