+++ main/urls.py
@@ -1,9 +1,7 @@
 from django.conf import settings
 from django.conf.urls.static import static
 from django.contrib import admin
-from django.urls import include, path, reverse
-from django.views.generic.base import RedirectView
-
+from django.urls import include, path
 
 from apps.homepage import views
 
@@ -21,21 +19,12 @@ urlpatterns = [
     path('jobs/', include('apps.jobs.urls', namespace='jobs')),
 ]
 
-# redirections
-urlpatterns += [
-    path(
-        'join',
-        RedirectView.as_view(url=reverse('members:join')),
-        name='goto_join_members',
-    )
-]
-
 if settings.DEBUG:
     import debug_toolbar
 
     urlpatterns.append(
         path('__debug__/', include(debug_toolbar.urls)),
-        )
+    )
     urlpatterns += static(
         settings.MEDIA_URL, document_root=settings.MEDIA_ROOT
     )
